Add gitignore origin for each file to output
In addition to the file name and size, this commit adds a third column
to the output that, for each file, shows why it is ignored.

In order to accomplish this, the code was restructured to feed the list
of ignored files to two separate commands for 1) computing the file size
and 2) obtaining the gitignore origin.

Due to this refactoring, the previous xargs-related switch for macOS
systems can be safely removed, because the file list can now be checked
for emptiness before running the subsequent commands (instead of having
to resort to the “-r” option of xargs).